Sidebar with Search
Made a lot of changes that are still in progress, but the sidebar is now always visible as well as the search bar. All that changes is the "view" which is either a playlist component or a search component. I would like to add routes to this so that if you are in a playlist view the adress bar reflects that and likewise with search.